Support queries that use $partition column for iceberg tables.
There are couple of known issues with current implementation. One is the classloader heck which should not be necessary given presto-spi would
invoke the system table methods with the same classloader context but it does not do that currently so we have to set the classloader correctly
on method invocations.

Iceberg API does not handle partitions with special characters like ' ' and the API returns a list of strings as partition names. This requires
presto to parse the partition path which would only work if iceberg escapes the = and / character. Ideally the iceberg API will give us a structure
that has partition column and value.