Handle the hidden partitioning case for the read path
When hidden partitioning is used the partitionData object we get from iceberg split is only accessible with indexes and not with ids.
On the split manager side and getColumnMetadata call we mark identity partition column as partition columns and ignore other partition columns as
we only want users to project/filter on actual source columns and to support the migrated tables which do not store the partition column's data.
This becomes a problem for hidden parition column tables that also have identity partitions i.e. for a table with to_hour(ts), x, to_day(ts) partitioning
scheme we only recognize x as partition column and retrieve the value of to_hour(ts) as the value of x as it is on index 0. This commit handles the case
by correctly getting the right partition data for identity columns.