Addressing Presto write side bug. We need to make the appendPage method thread safe and we need to ensure the state is property maintained when appendPage is called multiple times with different pages.
There is still more work to do on the write side to ensure we are not opening too many writers and to see if there is a
way to reduce small files issue. David during the review also recommended that we should consider moving to PageIndexer
which is something we should look into.